[refactor]:navbar move to newfile
