+++ src/components/index.js
@@ -1,3 +1,3 @@
-import NavBar from "./navbar";
+import NavBar from "./navbar/navbar";
 import { Sidebar } from "./sidebar/sidebar";
 export{NavBar,Sidebar}
